refactor: movel types for each trait to a .model.ts file
