+++ libs/ngrx-traits/signals/src/lib/index.ts
@@ -1,20 +1,28 @@
-export * from './util';
 export * from './with-call-status/with-call-status';
+export * from './with-call-status/with-call-status.model';
 export * from './with-entities-filter/with-entities-local-filter';
+export * from './with-entities-filter/with-entities-local-filter.model';
 export * from './with-entities-filter/with-entities-remote-filter';
 export * from './with-entities-pagination/with-entities-local-pagination';
+export * from './with-entities-pagination/with-entities-local-pagination.model';
 export * from './with-entities-pagination/with-entities-remote-pagination';
+export * from './with-entities-pagination/with-entities-remote-pagination.model';
 export * from './with-entities-pagination/with-entities-remote-scroll-pagination';
+export * from './with-entities-pagination/with-entities-remote-scroll-pagination.model';
 export * from './with-entities-pagination/signal-infinite-datasource';
 export {
   Sort,
   SortDirection,
 } from './with-entities-sort/with-entities-sort.utils';
 export * from './with-entities-sort/with-entities-local-sort';
+export * from './with-entities-sort/with-entities-local-sort.model';
 export * from './with-entities-sort/with-entities-remote-sort';
 export * from './with-entities-selection/with-entities-single-selection';
+export * from './with-entities-selection/with-entities-single-selection.model';
 export * from './with-entities-selection/with-entities-multi-selection';
+export * from './with-entities-selection/with-entities-multi-selection.model';
 export * from './with-entities-loading-call/with-entities-loading-call';
 export * from './with-logger/with-logger';
 export * from './with-calls/with-calls';
+export * from './with-calls/with-calls.model';
 export * from './with-sync-to-web-storage/with-sync-to-web-storage';
