+++ libs/ngrx-traits/signals/src/lib/with-entities-pagination/with-entities-infinite-pagination.util.ts
@@ -1,4 +1,4 @@
-import { capitalize } from '@ngrx-traits/signals';
+import { capitalize } from '../util';
 
 export function getWithEntitiesInfinitePaginationKeys(config?: {
   collection?: string;
