+++ libs/ngrx-traits/signals/src/lib/with-entities-pagination/with-entities-local-pagination.ts
@@ -15,55 +15,15 @@ import {
 import type { StateSignal } from '@ngrx/signals/src/state-signal';
 
 import { combineFunctions, getWithEntitiesKeys } from '../util';
+import {
+  EntitiesPaginationLocalComputed,
+  EntitiesPaginationLocalMethods,
+  EntitiesPaginationLocalState,
+  NamedEntitiesPaginationLocalComputed,
+  NamedEntitiesPaginationLocalMethods,
+  NamedEntitiesPaginationLocalState,
+} from './with-entities-local-pagination.model';
 import { getWithEntitiesLocalPaginationKeys } from './with-entities-local-pagination.util';
-
-export type EntitiesPaginationLocalState = {
-  entitiesPagination: {
-    currentPage: number;
-    pageSize: number;
-  };
-};
-export type NamedEntitiesPaginationLocalState<Collection extends string> = {
-  [K in Collection as `${K}Pagination`]: {
-    currentPage: number;
-    pageSize: number;
-  };
-};
-
-export type EntitiesPaginationLocalComputed<Entity> = {
-  entitiesCurrentPage: Signal<{
-    entities: Entity[];
-    pageIndex: number;
-    total: number | undefined;
-    pageSize: number;
-    pagesCount: number | undefined;
-    hasPrevious: boolean;
-    hasNext: boolean;
-  }>;
-};
-export type NamedEntitiesPaginationLocalComputed<
-  Entity,
-  Collection extends string,
-> = {
-  [K in Collection as `${K}CurrentPage`]: Signal<{
-    entities: Entity[];
-    pageIndex: number;
-    total: number | undefined;
-    pageSize: number;
-    pagesCount: number | undefined;
-    hasPrevious: boolean;
-    hasNext: boolean;
-  }>;
-};
-
-export type EntitiesPaginationLocalMethods = {
-  loadEntitiesPage: (options: { pageIndex: number }) => void;
-};
-export type NamedEntitiesPaginationLocalMethods<Collection extends string> = {
-  [K in Collection as `load${Capitalize<string & K>}Page`]: (options: {
-    pageIndex: number;
-  }) => void;
-};
 
 /**
  * Generates necessary state, computed and methods for local pagination of entities in the store.
