+++ libs/ngrx-traits/signals/src/lib/with-entities-selection/with-entities-multi-selection.ts
@@ -20,60 +20,14 @@ import { capitalize, combineFunctions, getWithEntitiesKeys } from '../util';
 import { getWithEntitiesFilterKeys } from '../with-entities-filter/with-entities-filter.util';
 import { getWithEntitiesLocalPaginationKeys } from '../with-entities-pagination/with-entities-local-pagination.util';
 import { getWithEntitiesSortKeys } from '../with-entities-sort/with-entities-sort.util';
-
-export type EntitiesMultiSelectionState = {
-  entitiesSelectedIdsMap: Record<string | number, boolean>;
-};
-export type NamedEntitiesMultiSelectionState<Collection extends string> = {
-  [K in Collection as `${K}SelectedIdsMap`]: Record<string | number, boolean>;
-};
-
-export type EntitiesMultiSelectionComputed<Entity> = {
-  entitiesSelected: Signal<Entity[]>;
-  isAllEntitiesSelected: Signal<'all' | 'none' | 'some'>;
-};
-export type NamedEntitiesMultiSelectionComputed<
-  Entity,
-  Collection extends string,
-> = {
-  [K in Collection as `${K}SelectedEntities`]: Signal<Entity[]>;
-} & {
-  [K in Collection as `isAll${Capitalize<string & K>}Selected`]: Signal<
-    'all' | 'none' | 'some'
-  >;
-};
-
-export type EntitiesMultiSelectionMethods = {
-  selectEntities: (
-    options: { id: string | number } | { ids: (string | number)[] },
-  ) => void;
-  deselectEntities: (
-    options: { id: string | number } | { ids: (string | number)[] },
-  ) => void;
-  toggleSelectEntities: (
-    options: { id: string | number } | { ids: (string | number)[] },
-  ) => void;
-  toggleSelectAllEntities: () => void;
-  clearEntitiesSelection: () => void;
-};
-
-export type NamedEntitiesMultiSelectionMethods<Collection extends string> = {
-  [K in Collection as `select${Capitalize<string & K>}Entities`]: (
-    options: { id: string | number } | { ids: (string | number)[] },
-  ) => void;
-} & {
-  [K in Collection as `deselect${Capitalize<string & K>}Entities`]: (
-    options: { id: string | number } | { ids: (string | number)[] },
-  ) => void;
-} & {
-  [K in Collection as `toggleSelect${Capitalize<string & K>}Entities`]: (
-    options: { id: string | number } | { ids: (string | number)[] },
-  ) => void;
-} & {
-  [K in Collection as `toggleSelectAll${Capitalize<
-    string & K
-  >}Entities`]: () => void;
-};
+import {
+  EntitiesMultiSelectionComputed,
+  EntitiesMultiSelectionMethods,
+  EntitiesMultiSelectionState,
+  NamedEntitiesMultiSelectionComputed,
+  NamedEntitiesMultiSelectionMethods,
+  NamedEntitiesMultiSelectionState,
+} from './with-entities-multi-selection.model';
 
 function getEntitiesMultiSelectionKeys(config?: { collection?: string }) {
   const collection = config?.collection;
