+++ libs/ngrx-traits/signals/src/lib/with-entities-selection/with-entities-single-selection.ts
@@ -16,41 +16,14 @@ import {
 import type { StateSignal } from '@ngrx/signals/src/state-signal';
 
 import { capitalize, combineFunctions, getWithEntitiesKeys } from '../util';
-
-export type EntitiesSingleSelectionState = {
-  entitiesSelectedId?: string | number;
-};
-export type NamedEntitiesSingleSelectionState<Collection extends string> = {
-  [K in Collection as `${K}SelectedId`]?: string | number;
-};
-
-export type EntitiesSingleSelectionComputed<Entity> = {
-  entitiesSelectedEntity: Signal<Entity | undefined>;
-};
-export type NamedEntitiesSingleSelectionComputed<
-  Entity,
-  Collection extends string,
-> = {
-  [K in Collection as `${K}SelectedEntity`]: Signal<Entity | undefined>;
-};
-export type EntitiesSingleSelectionMethods = {
-  selectEntity: (options: { id: string | number }) => void;
-  deselectEntity: (options: { id: string | number }) => void;
-  toggleEntity: (options: { id: string | number }) => void;
-};
-export type NamedEntitiesSingleSelectionMethods<Collection extends string> = {
-  [K in Collection as `select${Capitalize<string & K>}Entity`]: (options: {
-    id: string | number;
-  }) => void;
-} & {
-  [K in Collection as `deselect${Capitalize<string & K>}Entity`]: (options: {
-    id: string | number;
-  }) => void;
-} & {
-  [K in Collection as `toggle${Capitalize<string & K>}Entity`]: (options: {
-    id: string | number;
-  }) => void;
-};
+import {
+  EntitiesSingleSelectionComputed,
+  EntitiesSingleSelectionMethods,
+  EntitiesSingleSelectionState,
+  NamedEntitiesSingleSelectionComputed,
+  NamedEntitiesSingleSelectionMethods,
+  NamedEntitiesSingleSelectionState,
+} from './with-entities-single-selection.model';
 
 function getEntitiesSingleSelectionKeys(config?: { collection?: string }) {
   const collection = config?.collection;
